Adds first iteration of basic biorhythm formula. Captures user dob and tob using Dialog Fragments and stores using internal storage

+++ app/src/main/java/alexbrown/x/biorhythms/CalculationResults.kt
@@ -1,7 +1,9 @@
 package alexbrown.x.biorhythms
 
+import java.math.BigDecimal
+
 data class CalculationResults(
-    val physicalScore: Double,
-    val emotionalScore: Double,
-    val intellectualScore: Double
+    val physicalScore: BigDecimal,
+    val emotionalScore: BigDecimal,
+    val intellectualScore: BigDecimal
 )
